requirements.txt
asgiref==3.11.0
Django==6.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
# dotenv==0.9.9
PyJWT==2.10.1
python-dotenv==1.2.1
setuptools==80.9.0
sqlparse==0.5.5
tzdata==2025.3
wheel==0.45.1
